requirements.txt
openvino-dev
onnxruntime
numpy
pandas
streamlit  # Optional for UI later
